requirements.txt
chromadb
openai
youtube_transcript_api
ffmpeg-python
streamlit
python-dotenv
requests